Extract Green Lanes API key auth into concern
