+++ app/controllers/api/v2/green_lanes/base_controller.rb
@@ -4,6 +4,7 @@ module Api
       class BaseController < ApiController
         no_caching
 
+        include ApiKeyAuthenticatable
         include ApiTokenAuthenticatable
 
         before_action :check_service, :authenticate, :set_request_scope
@@ -29,42 +30,24 @@ module Api
         end
 
         def authenticate
-          unless Rails.env.development? && TradeTariffBackend.green_lanes_api_keys.blank?
-            authenticated = authenticate_with_api_keys
-          end
+          return if skip_authentication?
 
-          unless Rails.env.development? && TradeTariffBackend.api_tokens.blank?
-            authenticated ||= authenticate_with_api_tokens
-          end
+          authenticated = false
+          authenticated ||= authenticate_with_api_keys
+          authenticated ||= authenticate_with_api_tokens
 
-          unless authenticated || (Rails.env.development? && TradeTariffBackend.green_lanes_api_keys.blank?)
-            render json: { error: 'Invalid API Key' }, status: :bad_request
-          end
+          return if authenticated
+
+          render json: { error: 'Invalid API Key' }, status: :bad_request
         end
 
-        def authenticate_with_api_keys
-          provided_key = request.headers['X-Api-Key']
-          return false if provided_key.blank?
+        def skip_authentication?
+          return false unless Rails.env.development?
 
-          Rails.logger.debug "Provided key: #{provided_key}"
+          no_api_keys = api_keys.blank?
+          no_api_tokens = TradeTariffBackend.api_tokens.blank?
 
-          key, config = api_keys.find do |api_key, _config|
-            ActiveSupport::SecurityUtils.secure_compare(provided_key, api_key)
-          end
-
-          key.present?.tap do |authenticated|
-            @client_id = config['client_id'] if authenticated
-            @auth_type = 'x-api-key' if authenticated
-          end
-        end
-
-        def api_keys
-          @api_keys ||= read_api_keys
-        end
-
-        def read_api_keys
-          api_key_hash = JSON.parse(TradeTariffBackend.green_lanes_api_keys)
-          api_key_hash['api_keys']
+          no_api_keys && no_api_tokens
         end
       end
     end
